perf(auto-nvimrc): lazy-load sha2 library
Only require sha2 library when it is needed to compute the hash.

This means the library (which is over 100 kB) will not be loaded if
there are no nvimrc files to execute or the nvimrc files are marked as
"always" or "never" (their hash is not needed).

This should hopefully speed up the performance a bit.

The `compose` function is no longer needed. Dead code was removed