refactor(nvim): adjust default options
Remove `secure` and `exrc` since they do not seem to have any effects.

Change the colorcolumn to appear 1 character after the textwidth. Makes
editing a bit nicer, since the column now creates a box whose inside can
be filled in (instead of being able to fill up-to, and including the
border).