fix(i3): enabling lock screen service on boot
betterlockscreen service could not be enabled on boot because the
symlinked service definition could not be found. This is possible
because it was a symlink to the user's home folder, which may not have
been loaded at the time of systemd initialization.

Copying the service instead of symlinking it seems to have solved the
problem.