feat(nvim): use nvim-scrollbar instead of nvim-scrollview
nvim-scrollbar can display diagnostics in the sidebar, which
nvim-scrollview does not. I have not been using it to drag the file
anyway, so the lack of mouse support is not a show-stopper.

The highlight for the sidebar is changed from WildMenu to Visual (a
darker background). This way the diagnostics stand out more.