feat(nvim): use nvim-lsp-installer to install LSP servers
Leverage nvim-lsp-installer to install LSP servers.

This allows removing the custom install scripts for the LSP servers.

The only LSP server install script left is the one for jdtls (Java),
because I do not use if often and it was difficult to set up.

All other scripts in the `lsp-servers` directory install utilities
useful when working with a given language. These also include CLI tools
used by null-ls.

All other LSP servers are installed and maintained using
nvim-lsp-installer. It makes it easy to install the servers using
the `:InstallDefaultLspServers` command and also update them inside the
`:LspInstallInfo` popup.