feat(i3): enable hidpi support
Increase the DPI to 2x and then use `xrandr --scale 1.5x1.5` to scale
down the UI. This results in sharper image when using 4k monitors
compared to switching them to 1440p resolution.

It worked well on my first 27 inch monitor (connected via HDMI). I could
not get my second monitor to work, though. The laptop is connected via
USB-C into an adapter, and the adapter is connected to the monitor via
HDMI. The monitor said it does not detect any signal.