fix(i3): initialize laptop screen after starting X server
The laptop screen is not automatically turned on after `startx`.
It needs to be manually turned on via `xrandr`.

It also turns out that running xrandr requires the X session to be
visible. Otherwise, the X server crashes for me. Thus, using `sleep 5 &&
xrandr ...` works if I switch to the X session before 5 seconds. I use
that as an escape hatch if something fails during X server
initialization.