refactor(nvim): extract auto-nvimrc to a separate repository
Extract the plugin to https://github.com/Gelio/auto-nvimrc.
This way I can iterate on the plugin in isolation without adding many
commits to my dotfiles. Also, other people can now use the plugin
(although it is not documented well yet and it is in an early stage).

I stopped depending on `sha2.lua` during the migration and used
`vim.fn.sha256` instead. It results in less Lua code added by the
plugin. The hashes are the same in both approaches.