fix(nvim): get rid of the annoying neo-tree cwd prompt
When the neo-tree's cwd did not include the current file, there was an
annoying "File not in cwd" prompt. I intentionally want to change
neo-tree's cwd to always be the parent directory of the current file.
The `reveal_force_cwd` flag prevents the prompt from appearing and does
what I want.