Revert "refactor(picom): use egl backend"
The egl backend does not seem to work when using my NVIDIA GPU
reverse-PRIME.

Starting the home-4k screenlayout and having
`prime-select nvidia` results in a black screen. When exiting (Win +
Escape, E), the background is shown correctly. It seems that picom has
trouble using the egl backend to composite windows, but when it is
closed on exit, some UI elements are shown.

Reverting to the glx backend fixes the problem. I cannot use the egl
backend with my NVIDIA GPU for now. I can use it for `prime-select
on-demand` without any issues.

This reverts commit e45b449a3ec53b7c0a277c5240ce061aa60eb99f.