refactor(nvim): migrate to lazy.nvim
Change the plugin manager from packer to lazy.nvim.
Mark more plugins to be lazy-loaded. This speeds up neovim's startup
time (measured with `:StartupTime<CR>`) from 75ms to 35ms.

This 35ms is not exactly accurate, as some plugins start on `VeryLazy`
or `BufRead` events. This means that the difference in startup time will
be smaller when opening files. Still, more plugins are not loaded by
default now, which should help with overall responsiveness of the
editor during a longer editing session.