fix(nvim): unconditionally register LSP keymaps
Due to dynamic capabilities registration, some capabilities may not be
available when the server attaches to a buffer and may be registered at
a later time. This happens with tsserver's code actions. In such a case,
we still want to use these keymaps.

Instead of waiting for these capabilities to be registered, let's
unconditionally register the keymaps so they are always available in
LSP-attached buffers. This is the simplest but not a robust solution.