Revert "refactor(nvim): switch to typescript-tools.nvim"
This reverts commit f5a74275b454af1ff0779fe562a78e9071a5f278.

typescript-tools.nvim does not play well with Splitgraph's .nvimrc.lua.
Calling `setup` in the .nvimrc.lua throws an error, since supposedly the
module is not loaded yet.

When .nvimrc.lua is disabled, the plugin still has trouble locating a
tsserver instance (it is in `src/js/node_modules`).

Using the previous plugin is more stable.