fix(nvim): initialize markdown-preview.nvim on file load
The MarkdownPreview command seems to be registered in an autocmd or an
ftplugin. It is not loaded automatically when the plugin itself it
loaded. Thus, I had to `:e` the document to get it to register.
Executing the command for the first time caused an error.

I believe it is acceptable to load markdown-preview.nvim for all
Markdown files.