refactor(nvim): use l to open items in neo-tree
Problem: The previous mapping (o) is now used to order tree items.
Trying to open a file introduces unnecessary lag because Neovim waits to
see if we want to open a file or order items.

Solution: Use another mapping (l) to open files. It does not collide
with any other existing useful mapping. It replaces the "focus preview"
mapping, which is not used, because previews are not allowed in Netrw
mode.