refactor(nvim): use typescript-tools.nvim instead of typescript.nvim
typescript.nvim has been deprecated. typescript-tools.nvim seems to be
the new maintained plugin for TypeScript. It does not require
typescript-language-server to be installed.

One missing piece that tsserver supports and typescript-tools.nvim does
not, is loading all projects in a given solution. In tsserver, changing
the `root_dir` to a directory that contains a tsconfig solution file
was enough for tsserver to load all TS projects from that solution file.
In typescript-tools.nvim, doing the same only loads the projects that
are visited.

There is an issue [0] about the general problem is showing references
from other TS projects.

[0]: https://github.com/pmizio/typescript-tools.nvim/issues/124