feat(nvim): enable css_variables language server
A new LSP server for working with CSS variables. It has better support
for CSS variables than cssls, because it detects CSS variables defined
in other CSS files.

This LSP server is not yet added to `mason-lspconfig`. There was a
warning printed when attempting to include it in the `ensure_installed`
table. Thus, this server should be filtered out of that table to avoid
this unnecessary warnings.