refactor(macos): use left Option to control yabai
Change the skhd config to use left Option (lalt) instead of `fn` to
control Yabai.

This makes it easier to use on a custom keyboard, because custom
keyboards to not have a good way of simulating the Fn key
cross-platform. I also don't want to be using the CapsLock key on my
custom keyboard.

I use Left Option very rarely, so this seems like a safe choice.