fix(macos): use ctrl as mouse modifier for yabai
I only had Alt mapped on the right side of the keyboard. This meant that
I had to move both hands to the right to use the mouse and resize or
move windows.

Using some of the left hand side modifiers (Ctrl, Shift, Cmd) should be
much easier to do. I started with Ctrl, let's see if that breaks
anything.