refactor(firenvim): enable plugin by default
Enable firenvim by default but avoid configuring it when Neovim was not
started by firenvim.

This way I can update firenvim as regular when using Lazy. Previously,
firenvim did not show up in Lazy because it was disabled using the
`cond`, and I had to remember to remove `cond` to update the plugin.