fix(nvim): only use clangd for C and C++ files
By the time `lsp/clangd.lua` run and changed the `filetypes` for which
it should register, clangd is already queued to start (because the
`lsp/clangd.lua` file is sourced as the server is starting). This is too
late to unregister the LSP client.

In this case, the LSP configuration needs to be done in the main part of
the configuration.